Some portability fixes
git-svn-id: file:///home/cbm/dev/rpgtoolkit/svn/trunk@1900 8ad37bdc-5837-4682-a52a-fbb6d9f1e9f7

+++ vc/trans3/movement/CPlayer/CPlayer.cpp
@@ -341,7 +341,7 @@ void CPlayer::calculateLevels(const bool init)
 	pLevels->reserve(numLevels);
 	pLevels->push_back(double(m_playerMem.stats.experience));
 
-	unsigned double exp = double(m_playerMem.levelType);
+	double exp = double(m_playerMem.levelType);
 	for (unsigned int i = 1; i < numLevels; ++i)
 	{
 		pLevels->push_back((*pLevels)[i - 1] + exp);
